Fix portfolio view

+++ backend/src/backend/schemas/portfolio.py
@@ -1,5 +1,3 @@
-from decimal import Decimal
-
 from pydantic import BaseModel, ConfigDict, Field
 
 
@@ -8,16 +6,16 @@ class PortfolioRowResponse(BaseModel):
 
     owner: str = Field(description="Portfolio owner name.")
     broker_platform: str = Field(description="Broker platform identifier (e.g. 'ibkr', 'n26').")
-    total_invested: Decimal = Field(description="Net capital deployed: Σ(buy qty*price) - Σ(sell qty*price).")
-    current_value: Decimal | None = Field(
+    total_invested: float = Field(description="Net capital deployed: Σ(buy qty*price) - Σ(sell qty*price).")
+    current_value: float | None = Field(
         default=None,
         description="Σ(net_quantity * latest_price). None when any held ISIN has no price record.",
     )
-    performance_abs: Decimal | None = Field(
+    performance_abs: float | None = Field(
         default=None,
         description="current_value - total_invested. None when current_value is None.",
     )
-    performance_pct: Decimal | None = Field(
+    performance_pct: float | None = Field(
         default=None,
         description="performance_abs / total_invested * 100. None when current_value is None.",
     )
